Access Token, user id's
Created userModel, that have uname,email and password, validation added. Installed bcrypt and used to hash the password for the user. Imported jsonWebToken which helps for the auth part, that is using the auth token. Every part of the CRUD operations is made private, the way you can access it is with that access token. User controller, have it's own validation part of, registering, logging and getting the information about the current user.

+++ ContactAPI/controllers/contactController.js
@@ -2,14 +2,14 @@ const asyncHandler = require("express-async-handler")
 const Contact = require("../models/contactModel")
 //@ desc Get all contacts
 //@ route GET /api/contacts
-//@ access public 
+//@ access private 
 const getContact = asyncHandler(async(req,res)=>{
-    const contact = await Contact.find()
+    const contact = await Contact.find({user_id: req.user.id})
     res.status(200).json(contact)
 })
 //@ desc Create contact
 //@ route POST /api/contacts
-//@ access public 
+//@ access priavte 
 const createContact = asyncHandler(async (req,res)=>{
     const {name,email,phone} = req.body
     console.log("The request body is :",req.body)
@@ -22,7 +22,8 @@ const createContact = asyncHandler(async (req,res)=>{
     const contact = await Contact.create({
         name,
         email,
-        phone
+        phone,
+        user_id: req.user.id
     })
 
     res.status(201).json(contact)
@@ -32,7 +33,7 @@ const createContact = asyncHandler(async (req,res)=>{
 
 //@ desc Get contact by id
 //@ route GET /api/contacts/:id
-//@ access public 
+//@ access private 
 const getContactById = asyncHandler(async(req,res)=>{
  
 
@@ -46,13 +47,18 @@ const getContactById = asyncHandler(async(req,res)=>{
 
 //@ desc Update contact by id
 //@ route PUT /api/contacts/:id
-//@ access public 
+//@ access private 
 const updateContact = asyncHandler(async(req,res)=>{
     
     const contact = await Contact.findById(req.params.id)
     if(!contact) {
         res.status(404)
         throw new Error("Content not found")
+    }
+
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("User don't have permission to update other user contacts")
     }
 
     const updatedContact = await Contact.findByIdAndUpdate(
@@ -66,12 +72,17 @@ const updateContact = asyncHandler(async(req,res)=>{
 
 //@ desc Delete contact by id
 //@ route DELETE /api/contacts/:id
-//@ access public 
+//@ access private 
 const deleteContact = asyncHandler(async(req,res)=>{
     const contact = await Contact.findById(req.params.id)
     if(!contact) {
         res.status(404)
         throw new Error("Content not found")
+    }
+    
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("User don't have permission to delete other user contacts")
     }
     await Contact.findByIdAndDelete(req.params.id)
     res.status(200).json(contact)
